requirements.txt
Django==2.2.16
django-extensions==3.1.1
djangorestframework==3.12.2
factory-boy==3.2.0
Faker==6.5.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
